% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim_opacity.R
\name{anim_opacity}
\alias{anim_opacity}
\title{Animate an element through an opacity change}
\usage{
anim_opacity(opacity = NULL, timing = NULL, initial = FALSE)
}
\arguments{
\item{opacity}{The opacity value of the element at the keyframe time (given
as the LHS value in the \code{\link[=anims]{anims()}} call).}

\item{timing}{The timing function to use for the movement to the new
position.}

\item{initial}{Should this opacity value be the initial opacity value of
the element? If so, use \code{TRUE} and any value provided to \code{opacity} will be
disregarded.}
}
\description{
With an \code{\link[=anims]{anims()}} call, itself passed as to any \code{anims} argument, the
\code{anim_opacity()} function can be used to express an animation where the
target element undergoes a change in opacity with time.
}
\examples{
# Basic animation of an element's
# opacity value (moving to a new
# `opacity` value of `0`)
SVG(width = 300, height = 300) \%>\%
  svg_rect(
    x = 50, y = 50,
    width = 50, height = 50,
    attrs = attrs_pres(
      stroke = "magenta",
      fill = "lightblue"
    ),
    anims = anims(
      2.0 ~ anim_opacity(opacity = 0)
    )
  )

}
