% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anims.R
\name{ease_in}
\alias{ease_in}
\title{Use a timing function for an 'easing in' animation}
\usage{
ease_in(power = "basic")
}
\arguments{
\item{power}{The preset to use for the easing in cubic bezier function.}
}
\description{
Use a timing function for an 'easing in' animation
}
