% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anims.R
\name{ease_in_out}
\alias{ease_in_out}
\title{Use a timing function for an 'easing in and out' animation}
\usage{
ease_in_out(power = "basic")
}
\arguments{
\item{power}{The preset to use for the easing in cubic bezier function.}
}
\description{
Use a timing function for an 'easing in and out' animation
}
