% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anims.R
\name{ease_out}
\alias{ease_out}
\title{Use an 'easing out' animation}
\usage{
ease_out(power = "basic")
}
\arguments{
\item{power}{The preset to use for the easing in cubic bezier function.}
}
\description{
Use an 'easing out' animation
}
