% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.svg}
\alias{knit_print.svg}
\title{Safely print any \code{svg} object in R Markdown}
\usage{
knit_print.svg(x, ...)
}
\arguments{
\item{x}{an \code{svg} object.}
}
\description{
This will transform \code{svg} objects to \code{knit_print} objects
}
\keyword{internal}
