% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anims.R
\name{linear}
\alias{linear}
\title{Use a linear movement for animation}
\usage{
linear()
}
\value{
A \code{linear} function call as a string for use as a CSS property.
}
\description{
The \code{linear()} function can be used as a value for the \code{easing_fn} argument,
which is available in every \verb{anim_*()} function (e.g., \code{\link[=anim_position]{anim_position()}}).
}
