% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anims.R
\name{step_end}
\alias{step_end}
\title{Use a 'step-end' animation}
\usage{
step_end()
}
\value{
A \code{step-end} function call as a string for use as a CSS property.
}
\description{
The \code{step_end()} function can be used as a value for the \code{easing_fn}
argument, which is available in every \verb{anim_*()} function (e.g.,
\code{\link[=anim_position]{anim_position()}}).
}
