% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anims.R
\name{step_start}
\alias{step_start}
\title{Use a 'step-start' animation}
\usage{
step_start()
}
\value{
A \code{step-start} function call as a string for use as a CSS property.
}
\description{
The \code{step_start()} function can be used as a value for the \code{easing_fn}
argument, which is available in every \verb{anim_*()} function (e.g.,
\code{\link[=anim_position]{anim_position()}}).
}
