% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVG.R
\name{SVG}
\alias{SVG}
\title{Create an `svg` object}
\usage{
SVG(width, height, title = NULL, desc = NULL)
}
\arguments{
\item{width, height}{The dimensions of the SVG in `px` units.}

\item{title}{The `<title>` tag for the finalized SVG.}

\item{desc}{The `<desc>` tag for the finalized SVG.}
}
\description{
The `SVG()` function is the entry point for building an SVG from the ground
up. We can provided predefined `height` and `width` attributes that define
the canvas size for the SVG. From here, we would want to use functions that
add elements to the SVG object (e.g., [svg_rect()], [svg_circle()], etc.) and
thus progressively build the graphic.
}
