% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{svg_circle}
\alias{svg_circle}
\title{Addition of a `circle` element}
\usage{
svg_circle(svg, x, y, diameter, attrs = list(), anims = list(),
  id = NULL)
}
\arguments{
\item{svg}{The `svg` object that is created using the [SVG()] function.}

\item{x, y}{The `x` and `y` positions of the center of the circle to be
drawn. The `x` and `y` values are relative to upper left of the SVG
drawing area.}

\item{diameter}{The diameter of the circle shape in units of `px`.}

\item{attrs}{A presentation attribute list. The helper function
[attrs_pres()] can help us easily generate this named list object. For the
most part, the list's names are the presentation attribute names and the
corresponding values are the matching attribute values.}

\item{id}{An optional ID value to give to the built tag. This is useful for
modifying this element in a later function call or for interacting with
CSS.}
}
\description{
The `svg_circle()` function adds a circle to an `svg` object. The position of
the circle is given by `x` and `y`, and this refers to the center point of
the point of the circle. The `diameter` of the circle is given in units of
`px`.
}
\examples{
# Create an SVG with a single
# circle element
SVG(width = 80, height = 80) \%>\%
  svg_circle(
    x = 30, y = 30,
    diameter = 40,
    attrs = list(fill = "red")
  )

}
