% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{svg_ellipse}
\alias{svg_ellipse}
\title{Addition of an \code{ellipse} element}
\usage{
svg_ellipse(svg, x, y, width, height, attrs = list(), anims = list(),
  id = NULL)
}
\arguments{
\item{svg}{The \code{svg} object that is created using the \code{\link[=SVG]{SVG()}} function.}

\item{x, y}{The \code{x} and \code{y} positions of the center of the ellipse to be
drawn. The \code{x} and \code{y} values are relative to upper left of the SVG drawing
area.}

\item{width, height}{The \code{width} and \code{height} of the ellipse that is to be
drawn. The \code{width} is the overall width of the ellipse in the 'x' direction,
centered on point \code{x}. The \code{height} is the distance in the 'y' direction,
centered on point \code{y}.}

\item{attrs}{A presentation attribute list. The helper function
\code{\link[=attrs_pres]{attrs_pres()}} can help us easily generate this named list object. For the
most part, the list's names are the presentation attribute names and the
corresponding values are the matching attribute values.}

\item{anims}{An animation directive list for the element. This should be
structured using the \code{\link[=anims]{anims()}} function.}

\item{id}{An optional ID value to give to the built tag. This is useful for
modifying this element in a later function call or for interacting with
CSS.}
}
\description{
The \code{svg_ellipse()} function adds an ellipse to an \code{svg} object. The position
of the ellipse is given by \code{x} and \code{y}, and they refer to the center point of
the point of the ellipse. The \code{width} and the \code{height}, both in units of
\code{px}, provide the horizontal and vertical extents of the ellipse.
}
\examples{
# Create an SVG with a single
# ellipse element
svg <-
  SVG(width = 60, height = 60) \%>\%
    svg_ellipse(
      x = 30, y = 30,
      width = 50, height = 30,
      attrs = attrs_pres(fill = "purple")
    )

}
