% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{svg_filter}
\alias{svg_filter}
\title{Build an SVG `<filter>`}
\usage{
svg_filter(svg, id, width = NULL, height = NULL, filters = list())
}
\arguments{
\item{svg}{The `svg` object that is created using the [SVG()] function.}

\item{id}{The ID value to assign to the filter. This must be provided and it
should be unique among all `<filter>` elements.}

\item{width, height}{The lengths of `width` and `height` define the extent of
the filter.}

\item{filters}{A list of `filter_*()` function calls. Examples include
[filter_image()] and [filter_gaussian_blur()].}
}
\description{
The `svg_filter()` let's us create a named `<filter>` element that we can
apply to any SVG elements (such as shapes). We can bundle one or more filter
elements by supplying a list of `filter_*()` calls to the `filters` argument.
}
