% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{svg_line}
\alias{svg_line}
\title{Addition of an `line` element}
\usage{
svg_line(svg, x1, y1, x2, y2, attrs = list(), anims = list(),
  id = NULL)
}
\arguments{
\item{svg}{The `svg` object that is created using the [SVG()] function.}

\item{x1, y1}{The `x` and `y` positions of the line's start point.}

\item{x2, y2}{The `x` and `y` positions of the line's end point.}

\item{attrs}{A presentation attribute list. The helper function
[attrs_pres()] can help us easily generate this named list object. For the
most part, the list's names are the presentation attribute names and the
corresponding values are the matching attribute values.}

\item{anims}{An animation directive list. Currently not implemented.}

\item{id}{An optional ID value to give to the built tag. This is useful for
modifying this element in a later function call or for interacting with
CSS.}
}
\description{
The `svg_line()` function adds a line to an `svg` object. The line is drawn
using a start point (`x1` and `y1`) and an end point (`x2` and `y2`) points.
These positions are in units of `px`.
}
\examples{
# Create an SVG with a single
# line element
SVG(width = 100, height = 50) \%>\%
  svg_line(
    x1 = 5, y1 = 5,
    x2 = 95, y2 = 45,
    attrs = list(stroke = "blue")
  )

}
