% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{svg_polyline}
\alias{svg_polyline}
\title{Addition of an `polyline` element}
\usage{
svg_polyline(svg, points, attrs = list(), anims = list(), id = NULL)
}
\arguments{
\item{svg}{The `svg` object that is created using the [SVG()] function.}

\item{points}{A numeric vector of points (with alternating values for `x` and
`y` positions) that define the polyline. This can also be a single-length
character vector that holds the formatted points string (space-separated
`x` and `y` values, and comma-separated points).}

\item{attrs}{A presentation attribute list. The helper function
[attrs_pres()] can help us easily generate this named list object. For the
most part, the list's names are the presentation attribute names and the
corresponding values are the matching attribute values.}

\item{id}{An optional ID value to give to the built tag. This is useful for
modifying this element in a later function call or for interacting with
CSS.}
}
\description{
The `svg_polyline()` function adds a polyline to an `svg` object. The
polyline is drawn by connecting a series of points with straight lines. The
points can be provided as a vector that's exactly divisible by two, or, as a
formatted string that adheres to the specification of the `points` attribute
of the SVG `<polyline>` tag. All point positions are in units of `px`.
}
\examples{
# Create an SVG with a single
# polyline element
SVG(width = 300, height = 300) \%>\%
  svg_polyline(
    points = c(
      10, 10, 15, 20, 20, 15, 25, 30, 30, 25,
      35, 40, 40, 35, 45, 50, 50, 45),
    attrs = list(
      stroke = "blue",
      fill = "none"
    )
  )

}
