% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{svg_text}
\alias{svg_text}
\title{Addition of an `text` element}
\usage{
svg_text(svg, x, y, text, path = NULL, attrs = list(),
  anims = list(), id = NULL)
}
\arguments{
\item{svg}{The `svg` object that is created using the [SVG()] function.}

\item{x, y}{The `x` and `y` positions of the upper left of the text to be
drawn. The `x` and `y` values are relative to upper left of the SVG drawing
area itself.}

\item{text}{A character vector that contains the text to be rendered.}

\item{path}{A single-length character vector that holds the formatted path
string.}

\item{attrs}{A presentation attribute list. The helper function
[attrs_pres()] can help us easily generate this named list object. For the
most part, the list's names are the presentation attribute names and the
corresponding values are the matching attribute values.}

\item{id}{An optional ID value to give to the built tag. This is useful for
modifying this element in a later function call or for interacting with
CSS.}
}
\description{
The `svg_text()` function adds text to an `svg` object. As with many of the
functions that create shape elements (such as [svg_rect()]), the starting
position is defined by `x` and `y` values. All point positions are in units
of `px`.
}
\examples{
# Create an SVG with a single
# text element
SVG(width = 300, height = 300) \%>\%
  svg_text(
    x = 10, y = 20,
    text = "A line of text"
  )

}
