% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim_rotation.R
\name{anim_rotation}
\alias{anim_rotation}
\title{Animate an element through rotation}
\usage{
anim_rotation(rotation = NULL, anchor = "center", timing = NULL,
  initial = FALSE)
}
\arguments{
\item{rotation}{The rotation value of the element at the keyframe time (given
as the LHS value in the \code{\link[=anims]{anims()}} call).}

\item{timing}{The timing function to use for the movement to the new
position.}

\item{initial}{Should this rotation value be the initial rotation state of
the element? If so, use \code{TRUE} and any value provided to \code{rotation} will be
disregarded.}
}
\description{
With an \code{\link[=anims]{anims()}} call, itself passed as to any \code{anims} argument, the
\code{anim_rotation()} function can be used to express an animation where the
target element undergoes a rotation change with time.
}
\examples{
# Basic animation of an element's
# rotation state (moving to a new
# `rotation` value)
SVG(width = 300, height = 300) \%>\%
  svg_rect(
    x = 50, y = 50,
    width = 50, height = 50,
    attrs = attrs_pres(
      stroke = "magenta",
      fill = "lightblue"
    ),
    anims = anims(
      2.0 ~ anim_rotation(rotation = 180)
    )
  )

}
