% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anims.R
\name{ease_in_out}
\alias{ease_in_out}
\title{Use an 'easing in and out' animation}
\usage{
ease_in_out(power = "basic")
}
\arguments{
\item{power}{The preset to use for the easing in cubic bezier function.}
}
\value{
A \code{cubic-bezier} function call as a string for use as a CSS property.
}
\description{
The \code{ease_in_out()} function can be used as a value for the \code{easing_fn}
argument, which is available in every \verb{anim_*()} function (e.g.,
\code{\link[=anim_position]{anim_position()}}).
}
