% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_gaussian_blur}
\alias{filter_gaussian_blur}
\title{Filter: add a gaussian blur to an element}
\usage{
filter_gaussian_blur(what = "source", stdev = 1)
}
\arguments{
\item{what}{What exactly should be blurred? By default, it is the \code{"source"}
image.}

\item{stdev}{The number of standard deviations for the blur effect.}
}
\description{
A gaussian blur effectively blurs an input image or shape by the amount
specified in \code{stdev}. The standard deviation of \code{stdev} is in direct
reference to the gaussian distribution that governs the extent of blurring.
}
\examples{
# Add a green ellipse to an SVG and
# then apply the `filter_gaussian_blur()`
# filter to blur the edges
SVG(width = 200, height = 100) \%>\%
  svg_filter(
    id = "blur",
    filters = list(
      filter_gaussian_blur(stdev = 2)
    )
  ) \%>\%
  svg_ellipse(
    x = 40, y = 40,
    width = 50, height = 30,
    attrs = attrs_pres(
      fill = "green",
      filter = "blur"
    )
  )

}
