% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVG.R
\name{SVG}
\alias{SVG}
\title{Create an \code{svg} object}
\usage{
SVG(width, height, title = NULL, desc = NULL, anim_iterations = "infinite")
}
\arguments{
\item{width, height}{The dimensions of the SVG in \code{px} units.}

\item{title}{The \verb{<title>} tag for the finalized SVG.}

\item{desc}{The \verb{<desc>} tag for the finalized SVG.}

\item{anim_iterations}{How many should an SVG animation (if defined by use of
the \code{\link[=anims]{anims()}} function) be played? By default this is \code{"infinite"} (i.e.,
looped indefinitely) but we can specify the animation iteration count as
a positive number.}
}
\description{
The \code{SVG()} function is the entry point for building an SVG from the ground
up. We can provide predefined \code{height} and \code{width} attributes that define
the canvas size for the SVG. From here, we would want to use functions that
add elements to the SVG object (e.g., \code{\link[=svg_rect]{svg_rect()}}, \code{\link[=svg_circle]{svg_circle()}}, etc.) and
thus progressively build the graphic.
}
\examples{
# Create an SVG with nothing drawn
# within it
svg <- SVG(width = 200, height = 100)

# Add a rectangle and then a circle
svg <-
  svg \%>\%
  svg_rect(
    x = 20, y = 20,
    width = 40, height = 40
  ) \%>\%
  svg_circle(
    x = 100, y = 40,
    diameter = 40
  )

}
