% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{svg_path}
\alias{svg_path}
\title{Addition of an `path` element}
\usage{
svg_path(svg, path, attrs = list(), anims = list(), id = NULL)
}
\arguments{
\item{svg}{The `svg` object that is created using the [SVG()] function.}

\item{path}{A single-length character vector that holds the formatted path
string.}

\item{attrs}{A presentation attribute list. The helper function
[attrs_pres()] can help us easily generate this named list object. For the
most part, the list's names are the presentation attribute names and the
corresponding values are the matching attribute values.}

\item{anims}{An animation directive list. Currently not implemented.}

\item{id}{An optional ID value to give to the built tag. This is useful for
modifying this element in a later function call or for interacting with
CSS.}
}
\description{
The `svg_path()` function adds a path to an `svg` object. A path can
potentially be quite complex (with an interplay of line and curve commands),
so, a hand-encoded `path` string is not often done by hand. For this reason,
the `path` argument accepts only a formatted string that complies with the
input requirements for the `d` attribute of the SVG `<path>` tag. All point
positions are in units of `px`.
}
\examples{
# Create an SVG with a single
# path element
SVG(width = 300, height = 300) \%>\%
  svg_path(
    path = "M 50 160 q 100 -300 200 0",
    attrs = list(
      stroke = "magenta",
      `stroke-width` = 5,
      fill = "lightblue"
    )
  )

}
